feat: implementando microsserviço pedido

+++ Ecommerce/Ecommerce.Commons/Dtos/PedidoDto.cs
@@ -8,7 +8,7 @@ namespace Ecommerce.Commons.Dtos
         public int IdUsuario { get; set; }
         public int? IdPagamento { get; set; }
         public StatusPedidoEnum StatusPedido { get; set; }
-        public long PrecoTotal { get; set; }
+        public double PrecoTotal { get; set; }
 
         public List<ProdutoPedidoDto> ProdutoPedido { get; set; }
     }
